Add CHANGELOG.md file, move 'GmailAPIWriteWrapper' helper class into 'GmailAPIModifyWrapper' in crud module
