+++ CHANGELOG.md
@@ -1,0 +1,53 @@
+# Gmail API Wrapper Changelog
+
+## Version Number [dd-mm-yyyy]
+---------------------------
+
+Features:
+
+- [x] Added this feature
+
+Bugs:
+
+- [x] Fixed that one
+
+Documentation:
+
+- [x] Updated these docs
+
+Breaking:
+
+- [x] This changes will break your code!!
+--------------------------------------------------------------------------------
+
+### 0.0.1a2 [05-09-2017]
+
+Bugs:
+
+- [x] Fix availbale LABELS checking in `GmailAPIReadWrapper` helper class
+
+Breaking:
+
+- [x] Renamed `GmailAPIWriteWrapper` to `GmailAPIModifyWrapper` and moved it
+    from `gmail_api_wrapper.crud.write` to `gmail_api_wrapper.crud.modify` module
+
+
+
+### 0.0.1a1 [04-09-2017]
+
+Features:
+
+- [x] Make `gmail_api_wrapper` a package
+- [x] Add `setup.py` file
+- [x] Add package `requirements/*`
+- [x] Add `GmailAPIConnection` helper class in `gmail_api_wrapper.connection.py` module
+- [x] Add `GmailAPIReadWrapper` helper class in `gmail_api_wrapper.crud.read.py` module
+- [x] Add `GmailAPIWriteWrapper` helper class in `gmail_api_wrapper.crud.modify.py` module
+
+
+### 0.0.1a0 [27-07-2017]
+
+Features:
+
+- [x] Project Initialization
+
