+++ gmail_api_wrapper/version.py
@@ -1,3 +1,3 @@
 """Version file."""
-__version__ = '0.0.1a1'
-version = (0, 0, 1, 'a', 1)
+__version__ = '0.0.1a2'
+version = (0, 0, 1, 'a', 2)
